refactor(view): 수익률 계산 로직 수정
0일때 0.00% 이렇게 출력하는 문제가 발생했습니다.

return 되는 값을 BigDecimal에서 double로 수정하여 정상적으로 0.0%로 출력되도록 수정하였습니다.

+++ src/main/java/lotto/global/util/YieldCalculator.java
@@ -4,13 +4,13 @@ import java.math.BigDecimal;
 import java.math.RoundingMode;
 
 public class YieldCalculator {
-    public static BigDecimal calculate(int purchasePrice, int winningPrice) {
+    public static double calculate(int purchasePrice, int winningPrice) {
         BigDecimal purchase = BigDecimal.valueOf(purchasePrice);
         BigDecimal winning = BigDecimal.valueOf(winningPrice);
         BigDecimal rate = winning
                 .divide(purchase, 10, RoundingMode.HALF_UP)
                 .multiply(BigDecimal.valueOf(100));
 
-        return rate.setScale(2, RoundingMode.HALF_UP);
+        return rate.setScale(2, RoundingMode.HALF_UP).doubleValue();
     }
 }
